Moved AsciiDataBase to Utilities and updated tests.
