+++ tests/ascii_data_temperature_time.prm
@@ -33,11 +33,8 @@ subsection Boundary temperature model
   subsection Ascii data model
     set Data file name       = box_2d_%s.%d.csv
     
-    set Number of x grid points = 3
-    set Number of y grid points = 3
     set First data file model time = 1e5
     set Data file time step = 2e5
-    set Scale factor = 1
   end
 end
